Testing (#57)
* Influence bar charts (#39)

* Designed Layouts for Bar Graph Dock Widget and added widgets to it

* Populated selectable actors based on DB Import, Turn Slider Synced,
Dimensions populated.

* Modifications Related to Dimensions

* Dummy Stacked Graph for testing purpose- In progress

* DB linked with Bars Module- Work in Progress

* Bar Charts synched with DB and actors, dimensions, turns

* Minor Changes

* Updated bin_width functionality,  Graph Names based on Dimension and turn
 Please enter the commit message for your changes. Lines starting

* dimension names now saved

There is now a table named DimensionDescription that stores the integer ID and a description name for each SMP dimension.  I have added a const to smp setting the max descrip to 256 chars

* Updated DB reading code to sync with test.db, Graphs title to sync with DB
Dimensions.

* Distinct colors for all actors in DB, colors for selected Actors and grey for unselected actors

* Fixed distinct colors for actors, auto scalable y-axis, separate control for individual actors and number of bars.

* Updated Bar Graph dimensions to change along with scenarios, Actor CheckBoxes are set with backgroup colors of Bar. yAxis Sync with Dimension changes.

* Removed unnecessary talking to console

In several places where data was logged to the SQLITE tables, there were unnecessary and virtually uninformative messages stating "Stored SQL for turn # of all estimators, actors, and positions". This was repeated many times in a block for each turn.  I commented out these lines of code.  It may be useful at some point in the future to turn it back on, but make the messages actually informative...

* Clarified selected bargain

At the end of the bargain resolution phase in each queue, it's stated  "Chosen bargain (...): #mMax/#nb" where mMax is of course 0-based, but np is the length. This was slightly confusing, so I replaced the "/#nb" with "out of #nb bargains" to make it more clear

* Minor aesthetic change to console output

Updated console message that says "Bargains involving actor #:..." to display actor id in padded 2 characters (as in a few places elsewhere) so the relevant block of text is aligned nicely.

* Line Plots feature Added to SMPQ (#44)

* Designed Layout and changes code for lineGraph

* Populated Actors, dimensions and Syncronised function calls of bar and
line graph

* Added controls to Actor CheckBox, Select All, Dimensions
ComboBox and Turn Slider to plot Graph and Synchronize with  changes in
values, Removed Negative and Fraction values on Horizontal and Vertical
Axis while scaling graph. Line Graph Plotting Techniques Modified to
eliminate Lag while plotting.

* Feature to Clear all Graph and Bar plottables and disconnect controls of
graph widgets, when DB is edited or Create csv, Edit/View CSV is selected.
Labels for line graph at Begging and ending of line (Paused).
Actor Name, Description and Influence is displayed as a tip message when
mouse is over a line in the graph.

* Feature to toggle ActorName Label when we click on line in line plot.
Clicking on lineGraph radio button clears all labels at once.

* Working on feature to plot spline, in lineplots(In Progress)

* changes in  plotting for turn 1 and turn of line plots

* Spline Interpolation to get smooth curves

* bargain table update logic for init and recv actors (#45)

* Correct Bargn table update

The code to update the Bargn table after resolution conditioned upon Turn_t, BargnId, Init_Act_i, and Recd_Act_j to get the correct row to update. However, when I reviewed the code, I observed it was not conditioning on ScenarioId, which means if multiple run with the same data were in the DB, they could potentially all be inadvertently updated.  I fixed this

* SMPQ updated to handle, Multi-Dimensional cumulative DB  (#56)

* Designed Layout and changes code for lineGraph

* Populated Actors, dimensions and Syncronised function calls of bar and
line graph

* Added controls to Actor CheckBox, Select All, Dimensions
ComboBox and Turn Slider to plot Graph and Synchronize with  changes in
values, Removed Negative and Fraction values on Horizontal and Vertical
Axis while scaling graph. Line Graph Plotting Techniques Modified to
eliminate Lag while plotting.

* Feature to Clear all Graph and Bar plottables and disconnect controls of
graph widgets, when DB is edited or Create csv, Edit/View CSV is selected.
Labels for line graph at Begging and ending of line (Paused).
Actor Name, Description and Influence is displayed as a tip message when
mouse is over a line in the graph.

* Feature to toggle ActorName Label when we click on line in line plot.
Clicking on lineGraph radio button clears all labels at once.

* Working on feature to plot spline, in lineplots(In Progress)

* changes in  plotting for turn 1 and turn of line plots

* Spline Interpolation to get smooth curves

* Code Updated to work for MultiDimension Cumulative DB.

* Updated code to work with MultiDimension, MultiScenario DB

* chan